make frontend for advanced search work
