+++ frontend/src/components/artwork/Artworks.tsx
@@ -3,7 +3,7 @@ import { getArtworksByCategory, useBatchDeleteArtworkMutation } from "../../api/
 import LoadingPage from "../../pages/LoadingPage"
 import React, { useMemo, useState } from "react"
 import Navbar from "../navbar/Navbar"
-import { useNavigate, useParams } from "react-router-dom"
+import { useLocation, useNavigate, useParams} from "react-router-dom"
 import SearchComponent from "../search/SearchComponent"
 import FileDropzone from "../FileDropzone"
 import { ReactComponent as PlusIcon } from "../../assets/icons/plus.svg"
@@ -25,6 +25,7 @@ const Artworks = () => {
     const [showWarningPopup, setShowWarningPopup] = useState(false)
     const [sortOrder, setSortOrder] = useState<string>("default")
     const [showEditCollection, setShowEditCollection] = useState<boolean>(false)
+    const location = useLocation()
 
     const [currentPage, setCurrentPage] = useState(1)
     const pageSize = 10
@@ -41,8 +42,8 @@ const Artworks = () => {
     ]
 
     const { data: artworkData } = useQuery({
-        queryKey: ["artwork", currentPage, pageSize],
-        queryFn: () => getArtworksByCategory(collection as string, currentPage, pageSize),
+        queryKey: ["artwork", currentPage, pageSize, location.search],
+        queryFn: () => getArtworksByCategory(collection as string, currentPage, pageSize, location.search),
         enabled: !!collection,
         keepPreviousData: true,
     })
